refactor utils for log
