+++ api/common/utils.js
@@ -1,5 +1,7 @@
 import BN from "bn.js";
 import { ZkWasmUtil } from "zkwasm-service-helper";
+import { testNets } from "./constants.js";
+import { logDivider } from "./log_utils.js";
 
 export function fromHexString(hexString) {
   hexString = hexString.startsWith("0x") ? hexString.slice(2) : hexString;
@@ -34,90 +36,12 @@ export function trimPrefix(str, prefix) {
   return str;
 }
 
-export function logDivider() {
-  const line = "=".repeat(process.stdout.columns);
-  console.log(line);
-}
-
-export function logLoadingAnimation() {
-  // If width is equal to process.stdout.columns, the bar will overflow into the next line.
-  // 4 is the length of the prefix "[*] ".
-  // 55 is about the same length as the longest message in this script.
-  const width = Math.min(process.stdout.columns - 4, 55);
-  let frame = 0;
-  let stop = false;
-
-  const frames = ["▓"];
-  let position = 0;
-  const intervalId = setInterval(() => {
-    if (stop) {
-      clearInterval(intervalId);
-      process.stdout.clearLine();
-      process.stdout.cursorTo(0);
-      return;
-    }
-
-    const currentFrame = frames[frame % frames.length];
-    const loadingBar = `[*] ${currentFrame.repeat(
-      position,
-    )}▒${currentFrame.repeat(width - position - 1)}`;
-
-    process.stdout.cursorTo(0);
-    process.stdout.write(loadingBar);
-
-    position = (position + 1) % width;
-
-    frame++;
-  }, 400);
-
-  return {
-    stopAndClear: () => {
-      stop = true;
-      process.stdout.clearLine();
-      process.stdout.cursorTo(0);
-    },
-  };
-}
-
 export function concatHexStrings(hexStrings) {
   let result = "";
   for (let hexString of hexStrings) {
     result += hexString.startsWith("0x") ? hexString.slice(2) : hexString;
   }
   return "0x" + result;
-}
-
-export function currentNpmScriptName() {
-  return process.env.npm_lifecycle_event;
-}
-
-export function logReceiptAndEvents(
-  rawreceiptList,
-  blockid,
-  matchedEventOffsets,
-  filteredEventList,
-) {
-  console.log(
-    "[*]",
-    rawreceiptList.length,
-    rawreceiptList.length > 1
-      ? "receipts fetched from block"
-      : "receipt fetched from block",
-    blockid,
-  );
-  console.log(
-    "[*]",
-    matchedEventOffsets.length / 7,
-    matchedEventOffsets.length / 7 > 1 ? "events matched" : "event matched",
-  );
-  for (let i in filteredEventList) {
-    for (let j in filteredEventList[i]) {
-      filteredEventList[i][j].prettyPrint(
-        "\tTx[" + i + "]Event[" + j + "]",
-        false,
-      );
-    }
-  }
 }
 
 // https://github.com/zkcrossteam/g1024/blob/916c489fefa65ce8d4ee1a387f2bd4a3dcca8337/src/utils/proof.ts#L7
@@ -148,4 +72,18 @@ export function parseArgs(raw) {
     }
   }
   return parsedInputs.flat();
+}
+
+export function getTargetNetwork(inputtedNetworkName) {
+  const validNetworkNames = testNets.map((net) => net.name.toLowerCase());
+  if (!validNetworkNames.includes(inputtedNetworkName.toLowerCase())) {
+    console.log(`[-] NETWORK NAME "${inputtedNetworkName}" IS INVALID.`, "\n");
+    console.log(`[*] Valid networks: ${validNetworkNames.join(", ")}.`, "\n");
+    logDivider();
+    process.exit(1);
+  }
+  const targetNetwork = testNets.find(
+    (net) => net.name.toLowerCase() === inputtedNetworkName.toLowerCase()
+  );
+  return targetNetwork;
 }
